change some color in the design, and organized the files
