+++ src/components/Schedule.css
@@ -1,5 +1,6 @@
 :root {
     --blood-red: #BE0000;
+    --red:#690202f8;
     --white: #ffffff;
     --black: #242424
 }
