+++ src/views/HeroSection.css
@@ -5,7 +5,7 @@
 }
 
 .hero-container {
-    background-image: url('/public/images/image-1.png');
+    background-image: url('../assets/image-1.png');
     background-size: cover;
     background-position: center;
     height: calc(100vh - 80px);
